feat: Generates a basic QDPX project with text sources and some codes
I have completed a basic set of functionality and I think the software is almost testable now. The app exports the QDPX file based on the repo data and the conversion actions defined. There are still issues with how saving and loading configurations are working (the coding by glob is not being loaded), but I'll fix those soon.